Bugfix/enable stricter compiler warnings cast qual write strings (#1385)
* Enable compiler warnings for stricter checks and remove unnecessary suppressions

* Apply const correctness across multiple files, including datalink interface name parameters across multiple files

* Apply const qualifications to the character buffer string functions.

* Apply const correctness to string parameters in bacapp.c and bacapp.h. Revise parsing of const strings using bacnet_stptok() to tokenize.

* Add bacnet_strncpy to ensure null-terminated partial string copying. Replace strncpy with bacnet_strncpy for safer string copying in modbus_rtu and point_table modules.

* Fix file compile warnings where the "no previous declaration" is being generated.

* Fix floating-point arithmetic by adding type casts for clarity, including a changing color_rgb_clamp to use float instead of double to fix compiler float conversion warnings.

* Suppress warnings for discarded qualifiers in CMake and Makefiles across multiple ports

* Refactor hostname validation and encoding functions for const correctness and improved string handling

* Add tests for bacapp_parse_application_data and weekly schedule parsing

* Rename routed_get_my_address to Routed_Device_Get_My_Address and update references for consistency with BAC_ROUTING

* Refactor Ringbuf functions to remove unnecessary volatile qualifiers and casting to improve type safety

* Change parameter type to const char * in bacapp_parse_application_data to fix compiler error.

* Move event detection function definitions and remove the redundant declarations.

* Fix warning about Device_Configuration_File_Encode in device.c

* Add conditional compilation for backup and restore functionality in device.c

* Fix cast to uintptr_t for interface name assignment in dlmstp.c and ztest_mock.h

* Fix variable qualifier in bip_get_local_netmask and suppress additional compiler warnings

* Enable additional compiler warnings in unit tests and treat them as errors to enforce clean builds

* Add parse_primitive_value function to handle primitive data parsing

* Add test case for Ethernet datalink in ZTEST framework

+++ src/bacnet/basic/sys/color_rgb.h
@@ -17,7 +17,7 @@ extern "C" {
 #endif /* __cplusplus */
 
 BACNET_STACK_EXPORT
-double color_rgb_clamp(double d, double min, double max);
+float color_rgb_clamp(float d, float min, float max);
 
 BACNET_STACK_EXPORT
 void color_rgb_to_xy(
